refactor: product 관련 기능들은 product 패키지로 분리
