+++ src/main/java/cart/auth/config/SecurityConfig.java
@@ -1,11 +1,8 @@
 package cart.auth.config;
 
-import cart.auth.filter.AuthenticationFilter;
 import cart.auth.resolver.AuthenticationResolver;
 import cart.repository.user.UserRepository;
 import java.util.List;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -19,13 +16,13 @@ public class SecurityConfig implements WebMvcConfigurer {
         this.userRepository = userRepository;
     }
 
-    @Bean
-    public FilterRegistrationBean<AuthenticationFilter> filterRegistrationBean() {
-        final FilterRegistrationBean<AuthenticationFilter> registrationBean = new FilterRegistrationBean<>();
-        registrationBean.setFilter(new AuthenticationFilter(userRepository));
-        registrationBean.addUrlPatterns("/*");
-        return registrationBean;
-    }
+//    @Bean
+//    public FilterRegistrationBean<AuthenticationFilter> filterRegistrationBean() {
+//        final FilterRegistrationBean<AuthenticationFilter> registrationBean = new FilterRegistrationBean<>();
+//        registrationBean.setFilter(new AuthenticationFilter(userRepository));
+//        registrationBean.addUrlPatterns("/*");
+//        return registrationBean;
+//    }
 
     @Override
     public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> argumentResolvers) {
