+++ src/main/java/cart/controller/product/ProductRestController.java
@@ -1,10 +1,10 @@
-package cart.controller;
+package cart.controller.product;
 
 import cart.domain.product.Product;
 import cart.dto.ProductRequest;
 import cart.dto.ProductResponse;
-import cart.service.ProductCommandService;
-import cart.service.ProductQueryService;
+import cart.service.product.ProductCommandService;
+import cart.service.product.ProductQueryService;
 import java.net.URI;
 import java.util.List;
 import java.util.stream.Collectors;
