+++ src/test/java/cart/controller/AbstractControllerTest.java
@@ -1,8 +1,8 @@
 package cart.controller;
 
 import cart.repository.user.UserRepository;
-import cart.service.ProductCommandService;
-import cart.service.ProductQueryService;
+import cart.service.product.ProductCommandService;
+import cart.service.product.ProductQueryService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -10,7 +10,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest
-public abstract class AbstractProductControllerTest {
+public abstract class AbstractControllerTest {
 
     @MockBean
     protected ProductQueryService productQueryService;
